// apps/dashboard/src/api/generated/restaurantDashboardAPI.schemas.ts
export interface Customer {
  id: string;
  /** @maxLength 200 */
  name: string;
  /** @maxLength 320 */
  email: string;
  /** @maxLength 30 */
  phone: string;
  /**
     * @minimum -2147483648
     * @maximum 2147483647
     */
  totalVisits: number;
  totalSpent: string;
  lastVisit: string | null;
  tags: (string | null)[];
  notes: string | null;
  createdAt: string;
  updatedAt: string;
}